Created a backend folder for backend related code in order to organize the code
